Separate basic and composed selectors
