+++ src/state/simularium/selectors/basic.ts
@@ -1,7 +1,4 @@
-import { State } from "../types";
+import { State } from "../../types";
 
-// BASIC SELECTORS
 export const getSimulariumController = (state: State) =>
     state.simularium.simulariumController;
-
-// COMPOSED SELECTORS
